# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HKUST-NFT Server API",
        "title": "HKUST-NFT",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/events": {
            "get": {
                "description": "all events",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "responses": {
                    "200": {
                        "description": "GET/api/v1/events",
                        "schema": {
                            "$ref": "#/definitions/v1.EventsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
        "/events/{event-id}": {
            "get": {
                "description": "single event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "event id",
                        "name": "event-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET/api/v1/events/balala",
                        "schema": {
                            "$ref": "#/definitions/v1.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
        "/events/{event-id}/items": {
            "get": {
                "description": "items in event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "event id",
                        "name": "event-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET/api/v1/events/balala/items",
                        "schema": {
                            "$ref": "#/definitions/v1.EventItemsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
        "/events/{event-id}/join": {
            "post": {
                "description": "user join event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "event id",
                        "name": "event-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "POST/api/v1/events/balala/join",
                        "schema": {
                            "$ref": "#/definitions/v1.JoinEventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
        "/events/{event-id}/likes": {
            "post": {
                "description": "user join event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "event id",
                        "name": "event-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "item id",
                        "name": "item-id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "POST/api/v1/events/balala/likes",
                        "schema": {
                            "$ref": "#/definitions/v1.JoinEventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
        "/events/{event-id}/ranks": {
            "get": {
                "description": "item ranks in event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "event id",
                        "name": "event-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET/api/v1/events/balala/ranks",
                        "schema": {
                            "$ref": "#/definitions/v1.EventItemsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
        "/events/{event-id}/submit-item": {
            "post": {
                "description": "user submit item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "event id",
                        "name": "event-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "item id",
                        "name": "item-id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "POST/api/v1/events/balala/submit-item",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
        "/items": {
            "get": {
                "description": "all items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "favorites,popular,newest",
                        "name": "sort-by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "image,video,audio",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET/api/v1/items",
                        "schema": {
                            "$ref": "#/definitions/v1.AllItemsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
        "/items/{item-id}": {
            "get": {
                "description": "single item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "collection id",
                        "name": "item-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET/api/v1/items/yiiiiiii",
                        "schema": {
                            "$ref": "#/definitions/v1.Item"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "utils.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "USER_NAME_EXIST"
                },
                "message": {
                    "type": "string",
                    "example": "user name exist"
                }
            }
        },
        "v1.AllItemsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "success"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.Item"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "v1.Event": {
            "type": "object",
            "properties": {
                "event_description": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "event_name": {
                    "type": "string",
                    "example": "helloworld"
                },
                "start_time": {
                    "type": "string"
                },
                "time_duration": {
                    "type": "integer"
                }
            }
        },
        "v1.EventItem": {
            "type": "object",
            "properties": {
                "collection_id": {
                    "type": "integer",
                    "example": 5
                },
                "image": {
                    "type": "string"
                },
                "item_id": {
                    "type": "integer",
                    "example": 123455
                },
                "localFavorites": {
                    "type": "integer",
                    "example": 1
                },
                "ownerId": {
                    "type": "string",
                    "example": "mazhengwang-ust-hk"
                }
            }
        },
        "v1.EventItemsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "SUCCESS"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.EventItem"
                    }
                },
                "totel": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "v1.EventResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "event": {
                    "$ref": "#/definitions/v1.Event"
                }
            }
        },
        "v1.EventsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.Event"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "v1.Item": {
            "type": "object",
            "properties": {
                "collection_id": {
                    "type": "integer",
                    "example": 5
                },
                "createrId": {
                    "type": "string",
                    "example": "mazhengwang-ust-hk"
                },
                "favorites": {
                    "type": "integer",
                    "example": 1
                },
                "image": {
                    "type": "string"
                },
                "item_id": {
                    "type": "integer",
                    "example": 123455
                },
                "ownerId": {
                    "type": "string",
                    "example": "mazhengwang-ust-hk"
                }
            }
        },
        "v1.JoinEventResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "SUCCESS,USER_JOINED"
                }
            }
        }
    }
}